requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
jinja2
python-multipart